+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Test package for ezsql."""
+"""Test package for PyQueryX."""